utils/requirements.txt
Pillow>=9.0.0
boto3>=1.26.0